@@ -18,7 +18,7 @@
 		"issues": "https://github.com/oscarotero/imagecow/issues"
 	},
 	"require": {
-		"php": ">=5.3.0"
+		"php": ">=5.5"
 	},
 	"autoload": {
 		"psr-4": {
