@@ -1,11 +1,9 @@
 <?php
+
 namespace Imagecow;
 
 use Imagecow\Utils\Dimmensions;
 
-/**
- * Abstract core class extended by the available libraries (GD, Imagick)
- */
 class Image
 {
     const LIB_GD = 'Gd';
@@ -16,34 +14,21 @@ class Image
 
     protected $image;
     protected $filename;
+    protected $clientHints = [
+        'dpr' => null,
+        'viewport-width' => null,
+        'width' => null,
+    ];
 
     /**
-     * Static function to create a new Imagecow instance from an image file or string
-     *
-     * @param string $image   The name of the image file or binary string
-     * @param string $library The name of the image library to use (Gd or Imagick). If it's not defined, detects automatically the library to use.
-     *
-     * @return Image The Imagecow instance
-     */
-    public static function create($image, $library = null)
-    {
-        //check if it's a binary string
-        if (!ctype_print($image)) {
-            return static::createFromString($image, $library);
-        }
-
-        return static::createFromFile($image, $library);
-    }
-
-    /**
-     * Static function to create a new Imagecow instance from an image file
+     * Static function to create a new Imagecow instance from an image file.
      *
      * @param string $image   The path of the file
      * @param string $library The name of the image library to use (Gd or Imagick). If it's not defined, detects automatically the library to use.
      *
      * @return Image
      */
-    public static function createFromFile($image, $library = null)
+    public static function fromFile($image, $library = null)
     {
         $class = self::getLibraryClass($library);
 
@@ -51,109 +36,18 @@ class Image
     }
 
     /**
-     * Static function to create a new Imagecow instance from a binary string
+     * Static function to create a new Imagecow instance from a binary string.
      *
      * @param string $string  The string of the image
      * @param string $library The name of the image library to use (Gd or Imagick). If it's not defined, detects automatically the library to use.
      *
      * @return Image
      */
-    public static function createFromString($string, $library = null)
+    public static function fromString($string, $library = null)
     {
         $class = self::getLibraryClass($library);
 
         return new static($class::createFromString($string));
-    }
-
-    /**
-     * Static function to filter the transform operations according to client properties
-     * Useful to generate responsive images
-     *
-     * @param string $client_properties The cookie value generated by Imagecow.js scripts with the client dimmensions.
-     * @param string $operations        The operations to transform the image
-     *
-     * @return string The operations that matches with the client properties.
-     */
-    public static function getResponsiveOperations($client_properties, $operations)
-    {
-        list($width, $height, $speed) = explode(',', $client_properties);
-
-        $width = intval($width);
-        $height = intval($height);
-        $transform = array();
-
-        foreach (explode(';', $operations) as $operation) {
-            if (!preg_match('/^(.+):(.+)$/', $operation, $matches)) {
-                if (!empty($operation)) {
-                    $transform[] = $operation;
-                }
-                continue;
-            }
-
-            if (self::clientMatch($matches[1], $width, $height, $speed)) {
-                $transform[] = $matches[2];
-            }
-        }
-
-        return implode('|', $transform);
-    }
-
-    /**
-     * Check whether the client match with a selector
-     *
-     * @param string  $selector The operations selector
-     * @param integer $width    The client width
-     * @param integer $height   The client height
-     * @param string  $speed    The client speed
-     *
-     * @return boolean
-     */
-    private static function clientMatch($selector, $width, $height, $speed)
-    {
-        foreach (explode(',', $selector) as $rule) {
-            $rule = explode('=', $rule, 2);
-            $value = intval($rule[1]);
-
-            switch ($rule[0]) {
-                case 'max-width':
-                    if ($width > $value) {
-                        return false;
-                    }
-                    break;
-
-                case 'min-width':
-                    if ($width < $value) {
-                        return false;
-                    }
-                    break;
-
-                case 'width':
-                    if ($width != $value) {
-                        return false;
-                    }
-                    break;
-
-                case 'max-height':
-                    if ($height > $value) {
-                        return false;
-                    }
-                    break;
-
-                case 'min-height':
-                    if ($height < $value) {
-                        return false;
-                    }
-                    break;
-
-                case 'height':
-                    if ($height != $value) {
-                        return false;
-                    }
-                    break;
-            }
-        }
-
-        return true;
     }
 
     /**
@@ -172,18 +66,87 @@ class Image
         }
     }
 
-    public function getImage()
-    {
-        return $this->image->getImage();
-    }
+    /**
+     * Set the available client hints
+     * 
+     * @param array $clientHints
+     * 
+     * @return self
+     */
+    public function setClientHints(array $clientHints) {
+        $normalize = [];
 
-    public function setImage($image)
-    {
-        return $this->image->setImage($image);
+        foreach ($clientHints as $key => $value) {
+            $normalize[strtolower($key)] = is_null($value) ? null : (float) $value;
+        }
+
+        if (array_diff_key($normalize, $this->clientHints)) {
+            throw new \InvalidArgumentException('Invalid client hints');
+        }
+
+        $this->clientHints = array_replace($this->clientHints, $normalize);
+
+        return $this;
     }
 
     /**
-     * Inverts the image vertically
+     * Set a default background color used to fill in some transformation functions.
+     *
+     * @param array $background The color in rgb, for example: array(0,127,34)
+     *
+     * @return self
+     */
+    public function setBackground(array $background)
+    {
+        $this->image->setBackground($background);
+
+        return $this;
+    }
+
+    /**
+     * Define the image compression quality for jpg images.
+     *
+     * @param int $quality The quality (from 0 to 100)
+     *
+     * @deprecated Use quality instead
+     *
+     * @return self
+     */
+    public function setCompressionQuality($quality)
+    {
+        error_log('The method `setCompressionQuality()` is deprecated. Use `quality()` instead.');
+
+        return $this->quality($quality);
+    }
+
+    /**
+     * Get the fixed size according with the client hints
+     * 
+     * @param int $width
+     * @param int $height
+     * 
+     * @return array
+     */
+    private function calculateClientSize($width, $height)
+    {
+        if ($this->clientHints['width'] !== null && $this->clientHints['width'] < $width) {
+            return Dimmensions::getResizeDimmensions($width, $height, $this->clientHints['width'], null);
+        }
+
+        if ($this->clientHints['viewport-width'] !== null && $this->clientHints['viewport-width'] < $width) {
+            return Dimmensions::getResizeDimmensions($width, $height, $this->clientHints['viewport-width'], null);
+        }
+
+        if ($this->clientHints['dpr'] !== null) {
+            $width *= $this->clientHints['dpr'];
+            $height *= $this->clientHints['dpr'];
+        }
+
+        return [$width, $height];
+    }
+
+    /**
+     * Inverts the image vertically.
      *
      * @return self
      */
@@ -195,7 +158,7 @@ class Image
     }
 
     /**
-     * Inverts the image horizontally
+     * Inverts the image horizontally.
      *
      * @return self
      */
@@ -207,7 +170,7 @@ class Image
     }
 
     /**
-     * Saves the image in a file
+     * Saves the image in a file.
      *
      * @param string $filename Name of the file where the image will be saved. If it's not defined, The original file will be overwritten.
      *
@@ -221,7 +184,7 @@ class Image
     }
 
     /**
-     * Gets the image data in a string
+     * Gets the image data in a string.
      *
      * @return string The image data
      */
@@ -231,7 +194,7 @@ class Image
     }
 
     /**
-     * Gets the mime type of the image
+     * Gets the mime type of the image.
      *
      * @return string The mime type
      */
@@ -241,9 +204,9 @@ class Image
     }
 
     /**
-     * Gets the width of the image
+     * Gets the width of the image.
      *
-     * @return integer The width in pixels
+     * @return int The width in pixels
      */
     public function getWidth()
     {
@@ -251,9 +214,9 @@ class Image
     }
 
     /**
-     * Gets the height of the image
+     * Gets the height of the image.
      *
-     * @return integer The height in pixels
+     * @return int The height in pixels
      */
     public function getHeight()
     {
@@ -261,7 +224,7 @@ class Image
     }
 
     /**
-     * Converts the image to other format
+     * Converts the image to other format.
      *
      * @param string $format The new format: png, jpg, gif
      *
@@ -275,16 +238,15 @@ class Image
     }
 
     /**
-     * Resizes the image maintaining the proportion (A 800x600 image resized to 400x400 becomes to 400x300)
+     * Resizes the image maintaining the proportion (A 800x600 image resized to 400x400 becomes to 400x300).
      *
-     * @param integer|string $width   The max width of the image. It can be a number (pixels) or percentaje
-     * @param integer|string $height  The max height of the image. It can be a number (pixels) or percentaje
-     * @param boolean|null   $enlarge
-     * @param boolean        $cover
+     * @param int|string $width   The max width of the image. It can be a number (pixels) or percentaje
+     * @param int|string $height  The max height of the image. It can be a number (pixels) or percentaje
+     * @param bool       $cover
      *
      * @return self
      */
-    public function resize($width, $height = 0, $enlarge = false, $cover = false)
+    public function resize($width, $height = 0, $cover = false)
     {
         $imageWidth = $this->getWidth();
         $imageHeight = $this->getHeight();
@@ -293,8 +255,9 @@ class Image
         $height = Dimmensions::getIntegerValue($height, $imageHeight);
 
         list($width, $height) = Dimmensions::getResizeDimmensions($imageWidth, $imageHeight, $width, $height, $cover);
+        list($width, $height) = $this->calculateClientSize($width, $height);
 
-        if (($width === $imageWidth) || (!$enlarge && $width > $imageWidth)) {
+        if ($width >= $imageWidth) {
             return $this;
         }
 
@@ -304,12 +267,12 @@ class Image
     }
 
     /**
-     * Crops the image
+     * Crops the image.
      *
-     * @param integer|string $width  The new width of the image. It can be a number (pixels) or percentaje
-     * @param integer|string $height The new height of the image. It can be a number (pixels) or percentaje
-     * @param integer|string $x      The "x" position to crop. It can be number (pixels), percentaje, [left, center, right] or one of the Image::CROP_* constants
-     * @param integer|string $y      The "y" position to crop. It can be number (pixels), percentaje or [top, middle, bottom]
+     * @param int|string $width  The new width of the image. It can be a number (pixels) or percentaje
+     * @param int|string $height The new height of the image. It can be a number (pixels) or percentaje
+     * @param int|string $x      The "x" position to crop. It can be number (pixels), percentaje, [left, center, right] or one of the Image::CROP_* constants
+     * @param int|string $y      The "y" position to crop. It can be number (pixels), percentaje or [top, middle, bottom]
      *
      * @return self
      */
@@ -321,9 +284,11 @@ class Image
         $width = Dimmensions::getIntegerValue($width, $imageWidth);
         $height = Dimmensions::getIntegerValue($height, $imageHeight);
 
+        list($width, $height) = $this->calculateClientSize($width, $height);
+
         switch ($x) {
-            case Image::CROP_BALANCED:
-            case Image::CROP_ENTROPY:
+            case self::CROP_BALANCED:
+            case self::CROP_ENTROPY:
                 list($x, $y) = $this->image->getCropOffsets($width, $height, $x);
                 break;
         }
@@ -339,26 +304,25 @@ class Image
     /**
      * Adjust the image to the given dimmensions. Resizes and crops the image maintaining the proportions.
      *
-     * @param integer|string $width   The new width in number (pixels) or percentaje
-     * @param integer|string $height  The new height in number (pixels) or percentaje
-     * @param integer|string $x       The "x" position to crop. It can be number (pixels), percentaje, [left, center, right] or one of the Image::CROP_* constants
-     * @param integer|string $y       The "y" position to crop. It can be number (pixels), percentaje or [top, middle, bottom]
-     * @param boolean        $enlarge
+     * @param int|string $width   The new width in number (pixels) or percentaje
+     * @param int|string $height  The new height in number (pixels) or percentaje
+     * @param int|string $x       The "x" position to crop. It can be number (pixels), percentaje, [left, center, right] or one of the Image::CROP_* constants
+     * @param int|string $y       The "y" position to crop. It can be number (pixels), percentaje or [top, middle, bottom]
      *
      * @return self
      */
-    public function resizeCrop($width, $height, $x = 'center', $y = 'middle', $enlarge = false)
+    public function resizeCrop($width, $height, $x = 'center', $y = 'middle')
     {
-        $this->resize($width, $height, $enlarge, true);
+        $this->resize($width, $height, true);
         $this->crop($width, $height, $x, $y);
 
         return $this;
     }
 
     /**
-     * Rotates the image
+     * Rotates the image.
      *
-     * @param integer $angle Rotation angle in degrees (anticlockwise)
+     * @param int $angle Rotation angle in degrees (anticlockwise)
      *
      * @return self
      */
@@ -372,13 +336,13 @@ class Image
     }
 
     /**
-     * Define the image compression quality for jpg images
+     * Define the image compression quality for jpg images.
      *
-     * @param integer $quality The quality (from 0 to 100)
+     * @param int $quality The quality (from 0 to 100)
      *
      * @return self
      */
-    public function setCompressionQuality($quality)
+    public function quality($quality)
     {
         $quality = intval($quality);
 
@@ -394,22 +358,8 @@ class Image
     }
 
     /**
-     * Set a default background color used to fill in some transformation functions
-     *
-     * @param array $background The color in rgb, for example: array(0,127,34)
-     *
-     * @return self
-     */
-    public function setBackground(array $background)
-    {
-        $this->image->setBackground($background);
-
-        return $this;
-    }
-
-    /**
      * Reads the EXIF data from a JPEG and returns an associative array
-     * (requires the exif PHP extension enabled)
+     * (requires the exif PHP extension enabled).
      *
      * @param null|string $key
      *
@@ -429,39 +379,40 @@ class Image
     }
 
     /**
-     * Transform the image executing various operations of crop, resize, resizeCrop and format
+     * Transform the image executing various operations of crop, resize, resizeCrop and format.
      *
      * @param string $operations The string with all operations separated by "|".
      *
      * @return self
      */
-    public function transform($operations = '')
+    public function transform($operations = null)
     {
-        if (!$operations) {
-            return $this;
+        //No transform operations, resize to fix the client size
+        if (empty($operations)) {
+            return $this->resize($this->getWidth(), $this->getHeight());
         }
 
         $operations = self::parseOperations($operations);
 
         foreach ($operations as $operation) {
-            switch (strtolower($operation['function'])) {
+            switch ($operation['function']) {
                 case 'crop':
                 case 'resizecrop':
                     if (isset($operation['params'][2])) {
                         switch ($operation['params'][2]) {
                             case 'CROP_ENTROPY':
-                                $operation['params'][2] = Image::CROP_ENTROPY;
+                                $operation['params'][2] = self::CROP_ENTROPY;
                                 break;
 
                             case 'CROP_BALANCED':
-                                $operation['params'][2] = Image::CROP_BALANCED;
+                                $operation['params'][2] = self::CROP_BALANCED;
                                 break;
                         }
                     }
                     break;
             }
 
-            call_user_func_array(array($this, $operation['function']), $operation['params']);
+            call_user_func_array([$this, $operation['function']], $operation['params']);
         }
 
         return $this;
@@ -479,8 +430,22 @@ class Image
     }
 
     /**
+     * Returns the image as base64 url.
+     *
+     * @return string|null
+     */
+    public function base64()
+    {
+        if (($string = $this->getString()) && ($mimetype = $this->getMimeType())) {
+            $string = base64_encode($string);
+
+            return "data:{$mimetype};base64,{$string}";
+        }
+    }
+
+    /**
      * Auto-rotate the image according with its exif data
-     * Taken from: http://php.net/manual/en/function.exif-read-data.php#76964
+     * Taken from: http://php.net/manual/en/function.exif-read-data.php#76964.
      *
      * @return self
      */
@@ -520,11 +485,11 @@ class Image
     }
 
     /**
-     * Check whether the image is an animated gif
+     * Check whether the image is an animated gif.
      *
      * Copied from: https://github.com/Sybio/GifFrameExtractor/blob/master/src/GifFrameExtractor/GifFrameExtractor.php#L181
      *
-     * @return boolean
+     * @return bool
      */
     protected function isAnimatedGif()
     {
@@ -541,11 +506,11 @@ class Image
 
         fclose($fh);
 
-        return ($count > 1);
+        return $count > 1;
     }
 
     /**
-     * Converts a string with operations in an array
+     * Converts a string with operations in an array.
      *
      * @param string $operations The operations string
      *
@@ -553,29 +518,29 @@ class Image
      */
     private static function parseOperations($operations)
     {
-        $valid_operations = array('resize', 'resizeCrop', 'crop', 'format');
+        $valid_operations = ['resize', 'resizecrop', 'crop', 'format', 'quality'];
         $operations = explode('|', str_replace(' ', '', $operations));
-        $return = array();
+        $return = [];
 
         foreach ($operations as $operations) {
             $params = explode(',', $operations);
-            $function = trim(array_shift($params));
+            $function = strtolower(trim(array_shift($params)));
 
-            if (!in_array($function, $valid_operations)) {
+            if (!in_array($function, $valid_operations, true)) {
                 throw new ImageException("The transform function '{$function}' is not valid");
             }
 
-            $return[] = array(
+            $return[] = [
                 'function' => $function,
                 'params' => $params,
-            );
+            ];
         }
 
         return $return;
     }
 
     /**
-     * Checks the library to use and returns its class
+     * Checks the library to use and returns its class.
      *
      * @param string $library The library name (Gd, Imagick)
      *
@@ -583,7 +548,7 @@ class Image
      *
      * @return string
      */
-    public static function getLibraryClass($library)
+    private static function getLibraryClass($library)
     {
         if (!$library) {
             $library = Libs\Imagick::checkCompatibility() ? self::LIB_IMAGICK : self::LIB_GD;
