@@ -1,83 +1,84 @@
 <?php
+
 namespace Imagecow\Utils;
 
 use Imagecow\ImageException;
 
 /**
- * Usefull dimmensions calculations
+ * Usefull dimmensions calculations.
  */
 class Dimmensions
 {
-    protected static $positionsKeywords = array(
+    protected static $positionsKeywords = [
         'top' => '0%',
         'left' => '0%',
         'middle' => '50%',
         'center' => '50%',
         'right' => '100%',
         'bottom' => '100%',
-    );
+    ];
 
     /**
-     * Calculate the dimensions of a resize
+     * Calculate the dimensions of a resize.
      *
-     * @param integer $oldWidth
-     * @param integer $oldHeight
-     * @param integer $newWidth
-     * @param integer $newHeight
-     * @param boolean $cover
+     * @param int  $oldWidth
+     * @param int  $oldHeight
+     * @param int  $newWidth
+     * @param int  $newHeight
+     * @param bool $cover
      *
      * @return array [width, height]
      */
     public static function getResizeDimmensions($oldWidth, $oldHeight, $newWidth, $newHeight, $cover = false)
     {
         if (empty($newHeight)) {
-            return array((int) $newWidth, (int) ceil(($newWidth/$oldWidth) * $oldHeight));
+            return [(int) $newWidth, (int) ceil(($newWidth / $oldWidth) * $oldHeight)];
         }
 
         if (empty($newWidth)) {
-            return array((int) ceil(($newHeight/$oldHeight) * $oldWidth), (int) $newHeight);
+            return [(int) ceil(($newHeight / $oldHeight) * $oldWidth), (int) $newHeight];
         }
 
-        $scaleWidth = $newWidth/$oldWidth;
-        $scaleHeight = $newHeight/$oldHeight;
+        $scaleWidth = $newWidth / $oldWidth;
+        $scaleHeight = $newHeight / $oldHeight;
 
         if ($cover) {
             if ($scaleWidth > $scaleHeight) {
-                return array((int) $newWidth, (int) ceil($scaleWidth * $oldHeight));
+                return [(int) $newWidth, (int) ceil($scaleWidth * $oldHeight)];
             }
 
             if ($scaleWidth < $scaleHeight) {
-                return array((int) ceil($scaleHeight * $oldWidth), (int) $newHeight);
+                return [(int) ceil($scaleHeight * $oldWidth), (int) $newHeight];
             }
         } else {
             if ($scaleWidth < $scaleHeight) {
-                return array((int) $newWidth, (int) ceil($scaleWidth * $oldHeight));
+                return [(int) $newWidth, (int) ceil($scaleWidth * $oldHeight)];
             }
 
             if ($scaleWidth > $scaleHeight) {
-                return array((int) ceil($scaleHeight * $oldWidth), (int) $newHeight);
+                return [(int) ceil($scaleHeight * $oldWidth), (int) $newHeight];
             }
         }
 
         if ($scaleWidth < $scaleHeight || ($cover && $scaleWidth > $scaleHeight)) {
-            return array((int) $newWidth, (int) ceil($scaleWidth * $oldHeight));
+            return [(int) $newWidth, (int) ceil($scaleWidth * $oldHeight)];
         }
 
         if ($scaleWidth > $scaleHeight || ($cover && $scaleWidth < $scaleHeight)) {
-            return array((int) ceil($scaleHeight * $oldWidth), (int) $newHeight);
+            return [(int) ceil($scaleHeight * $oldWidth), (int) $newHeight];
         }
 
-        return array((int) $newWidth, (int) $newHeight);
+        return [(int) $newWidth, (int) $newHeight];
     }
 
     /**
-     * Calculate a dimension value
+     * Calculate a dimension value.
      *
-     * @param integer|string $value
-     * @param integer        $relatedValue
-     * @param boolean        $position
+     * @param int|string $value
+     * @param int        $relatedValue
+     * @param bool       $position
      *
-     * @return integer
+     * @return int
      */
     public static function getIntegerValue($value, $relatedValue, $position = false)
     {
@@ -93,11 +94,11 @@ class Dimmensions
     }
 
     /**
-     * Calculate a dimension value
+     * Calculate a dimension value.
      *
-     * @param integer|string $value
-     * @param integer        $relatedValue
-     * @param boolean        $position
+     * @param int|string $value
+     * @param int        $relatedValue
+     * @param bool       $position
      *
      * @return string
      */
@@ -119,13 +120,13 @@ class Dimmensions
     }
 
     /**
-     * Calculates the x/y position
+     * Calculates the x/y position.
      *
-     * @param string|integer|null $position
-     * @param integer             $newValue
-     * @param integer             $oldValue
+     * @param string|int|null $position
+     * @param int             $newValue
+     * @param int             $oldValue
      *
-     * @return integer
+     * @return int
      */
     public static function getPositionValue($position, $newValue, $oldValue)
     {
