@@ -1,22 +1,23 @@
 <?php
+
 namespace Imagecow\Libs;
 
 use Imagecow\ImageException;
 
 /**
- * Interface implemented by all libraries
+ * Interface implemented by all libraries.
  */
 interface LibInterface
 {
     /**
-     * Check if the library is installed
+     * Check if the library is installed.
      *
-     * @return boolean
+     * @return bool
      */
     public static function checkCompatibility();
 
     /**
-     * Create an instance from an image file
+     * Create an instance from an image file.
      *
      * @param string $filename Name of the file to load
      *
@@ -25,7 +26,7 @@ interface LibInterface
     public static function createFromFile($filename);
 
     /**
-     * Create an instance from an image string
+     * Create an instance from an image string.
      *
      * @param string $string The image content
      *
@@ -34,61 +35,61 @@ interface LibInterface
     public static function createFromString($string);
 
     /**
-     * Saves the image in a file
+     * Saves the image in a file.
      *
      * @param string $filename Name of the file where the image will be saved. If it's not defined, The original file will be overwritten.
      */
     public function save($filename);
 
     /**
-     * Gets the image data in a string
+     * Gets the image data in a string.
      *
      * @return string The image data
      */
     public function getString();
 
     /**
-     * Gets the mime type of the image
+     * Gets the mime type of the image.
      *
      * @return string The mime type
      */
     public function getMimeType();
 
     /**
-     * Gets the width of the image
+     * Gets the width of the image.
      *
-     * @return integer The width in pixels
+     * @return int The width in pixels
      */
     public function getWidth();
 
     /**
-     * Gets the height of the image
+     * Gets the height of the image.
      *
-     * @return integer The height in pixels
+     * @return int The height in pixels
      */
     public function getHeight();
 
     /**
-     * Converts the image to other format
+     * Converts the image to other format.
      *
      * @param string $format The new format: png, jpg, gif
      */
     public function format($format);
 
     /**
-     * Resizes the image maintaining the proportion (A 800x600 image resized to 400x400 becomes to 400x300)
+     * Resizes the image maintaining the proportion (A 800x600 image resized to 400x400 becomes to 400x300).
      *
-     * @param integer $width  The max width of the image
-     * @param integer $height The max height of the image
+     * @param int $width  The max width of the image
+     * @param int $height The max height of the image
      */
     public function resize($width, $height);
 
     /**
-     * Calculates automatically the x,y positions of a crop using a specific method
+     * Calculates automatically the x,y positions of a crop using a specific method.
      *
-     * @param integer $width  The new width of the image
-     * @param integer $height The new height of the image
-     * @param string  $method The method name (for example: "Entropy")
+     * @param int    $width  The new width of the image
+     * @param int    $height The new height of the image
+     * @param string $method The method name (for example: "Entropy")
      *
      * @throws ImageException if the method is not available
      *
@@ -97,50 +98,50 @@ interface LibInterface
     public function getCropOffsets($width, $height, $method);
 
     /**
-     * Crops the image
+     * Crops the image.
      *
-     * @param integer $width  The new width of the image
-     * @param integer $height The new height of the image
-     * @param integer $x      The "x" position where start to crop
-     * @param integer $y      The "y" position where start to crop
+     * @param int $width  The new width of the image
+     * @param int $height The new height of the image
+     * @param int $x      The "x" position where start to crop
+     * @param int $y      The "y" position where start to crop
      */
     public function crop($width, $height, $x, $y);
 
     /**
-     * Rotates the image
+     * Rotates the image.
      *
-     * @param integer $angle Rotation angle in degrees (anticlockwise)
+     * @param int $angle Rotation angle in degrees (anticlockwise)
      */
     public function rotate($angle);
 
     /**
-     * Inverts the image vertically
+     * Inverts the image vertically.
      */
     public function flip();
 
     /**
-     * Inverts the image horizontally
+     * Inverts the image horizontally.
      */
     public function flop();
 
     /**
-     * Define the image compression quality for jpg images
+     * Define the image compression quality for jpg images.
      *
-     * @param integer $quality The quality (from 0 to 100)
+     * @param int $quality The quality (from 0 to 100)
      */
     public function setCompressionQuality($quality);
 
     /**
-     * Set a default background color used to fill in some transformation functions
+     * Set a default background color used to fill in some transformation functions.
      *
-     * @param array $background The color in rgb, for example: array(0, 127, 34)
+     * @param array $background The color in rgb, for example: [0, 127, 34]
      */
     public function setBackground(array $background);
 
     /**
-     * Defines the image as an animated image
+     * Defines the image as an animated image.
      *
-     * @param boolean $animated
+     * @param bool $animated
      */
     public function setAnimated($animated);
 }
