@@ -1,18 +1,21 @@
 <?php
+
 namespace Imagecow\Utils;
 
 /**
- * Generic color conversions functions
+ * Generic color conversions functions.
  */
 class Color
 {
     /**
-     * Returns a YUV weighted greyscale value
+     * Returns a YUV weighted greyscale value.
      *
-     * @param  int $r
-     * @param  int $g
-     * @param  int $b
+     * @param int $r
+     * @param int $g
+     * @param int $b
+     *
      * @return int
+     *
      * @see http://en.wikipedia.org/wiki/YUV
      */
     public static function rgb2bw($r, $g, $b)
